refactor: reservationwaiting 패키지명 컨벤션에 맞게 수정
