+++ src/main/java/roomescape/domain/reservationwaiting/ReservationWaiting.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservatinWaiting;
+package roomescape.domain.reservationwaiting;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
