+++ src/main/java/roomescape/dto/reservation/MyReservationResponse.java
@@ -1,6 +1,6 @@
 package roomescape.dto.reservation;
 
-import roomescape.domain.reservatinWaiting.ReservationWaiting;
+import roomescape.domain.reservationwaiting.ReservationWaiting;
 import roomescape.domain.reservation.Reservation;
 import roomescape.dto.reservationtime.ReservationTimeResponse;
 import roomescape.dto.theme.ThemeResponse;
