+++ src/main/java/roomescape/dto/reservationWaiting/ReservationWaitingResponse.java
@@ -1,6 +1,6 @@
 package roomescape.dto.reservationWaiting;
 
-import roomescape.domain.reservatinWaiting.ReservationWaiting;
+import roomescape.domain.reservationwaiting.ReservationWaiting;
 import roomescape.dto.reservationtime.ReservationTimeResponse;
 import roomescape.dto.theme.ThemeResponse;
