+++ src/main/java/roomescape/dto/reservationWaiting/ReservationWaitingSequence.java
@@ -1,6 +1,6 @@
 package roomescape.dto.reservationWaiting;
 
-import roomescape.domain.reservatinWaiting.ReservationWaiting;
+import roomescape.domain.reservationwaiting.ReservationWaiting;
 
 public record ReservationWaitingSequence(ReservationWaiting reservationWaiting, Long sequence) {
 }
